Reverting encode_qr.py file move and name change
