+++ tests/test_seedqr.py
@@ -2,7 +2,7 @@ import os
 from embit import bip39
 from seedsigner.helpers.qr import QR
 from seedsigner.models.decode_qr import DecodeQR, DecodeQRStatus
-from seedsigner.helpers.qr_encoders import SeedQrEncoder, CompactSeedQrEncoder
+from seedsigner.models.encode_qr import SeedQrEncoder, CompactSeedQrEncoder
 from seedsigner.models.qr_type import QRType
